Moved SIS button to optimal position
Moved SIS button so that it is next to "Course Summary" text

Also changed the button to be an icon rather than text
