+++ static/js/redux/ui/containers/top_bar_advising_container.jsx
@@ -15,7 +15,6 @@ GNU General Public License for more details.
 import { connect } from 'react-redux';
 import TopBarAdvising from '../top_bar_advising';
 import { getCurrentSemester } from '../../reducers/root_reducer';
-import { triggerSISImportDataModal } from '../../actions/modal_actions';
 
 
 const mapStateToProps = state => ({
@@ -25,9 +24,6 @@ const mapStateToProps = state => ({
 
 const TopBarAdvisingContainer = connect(
   mapStateToProps,
-  {
-    triggerSISImportDataModal,
-  },
 )(TopBarAdvising);
 
 export default TopBarAdvisingContainer;
