+++ static/js/redux/ui/top_bar_advising.jsx
@@ -13,8 +13,6 @@ GNU General Public License for more details.
 */
 
 import React from 'react';
-import ReactTooltip from 'react-tooltip';
-import PropTypes from 'prop-types';
 import SocialProfileContainer from './containers/social_profile_container';
 import * as SemesterlyPropTypes from '../constants/semesterlyPropTypes';
 
@@ -80,30 +78,6 @@ class TopBarAdvising extends React.Component {
   }
 
   render() {
-    const SISImportDataModalButton = (
-      <div className="cal-btn-wrapper">
-        <button
-          onClick={() => this.props.triggerSISImportDataModal()}
-          className="import-data"
-          data-tip
-          data-for="mock-btn-tooltip"
-        >
-          {/* TODO: Move import data button (below) to optimal position */}
-          <div className="import-data">
-            <p>Import Data</p>
-          </div>
-        </button>
-        <ReactTooltip
-          id="mock-btn-tooltip"
-          class="tooltip"
-          type="dark"
-          place="bottom"
-          effect="solid"
-        >
-          <span>SIS Import Data Modal</span>
-        </ReactTooltip>
-      </div>
-    );
     return (
       <div className="top-bar">
         <a href="/" className="semesterly-name">
@@ -129,9 +103,6 @@ class TopBarAdvising extends React.Component {
           </div>
         </div>
         <SocialProfileContainer />
-        <span>
-          {SISImportDataModalButton}
-        </span>
         <div className="navicon" onClick={this.toggleComments}>
           <span />
           <span />
@@ -142,7 +113,6 @@ class TopBarAdvising extends React.Component {
 }
 
 TopBarAdvising.propTypes = {
-  triggerSISImportDataModal: PropTypes.func.isRequired,
   userInfo: SemesterlyPropTypes.userInfo.isRequired,
   currentSemester: SemesterlyPropTypes.semester.isRequired,
 };
